+++ lib/zoi/types/object.ex
@@ -1,7 +1,7 @@
 defmodule Zoi.Types.Object do
   @moduledoc false
 
-  use Zoi.Type.Def, fields: [:fields, :strict]
+  use Zoi.Type.Def, fields: [:fields, :strict, :coerce]
 
   def new(fields, opts) when is_map(fields) do
     apply_type(opts ++ [fields: fields])
@@ -27,7 +27,15 @@ defmodule Zoi.Types.Object do
       {:error, schema.meta.error || "invalid type: must be a map"}
     end
 
-    defp do_parse(%Zoi.Types.Object{fields: fields, strict: strict}, input, opts, path, errs) do
+    defp do_parse(
+           %Zoi.Types.Object{fields: fields, strict: strict, coerce: coerce},
+           input,
+           opts,
+           path,
+           errs
+         ) do
+      coerce = Keyword.get(opts, :coerce, coerce)
+
       unknown_fields_errors =
         if strict do
           unknown_fields(fields, input)
@@ -37,7 +45,7 @@ defmodule Zoi.Types.Object do
         |> Enum.map(&Zoi.Error.add_path(&1, path))
 
       Enum.reduce(fields, {%{}, errs, path}, fn {key, type}, {parsed, errors, path} ->
-        case map_fetch(input, key) do
+        case map_fetch(input, key, coerce) do
           :error ->
             cond do
               optional?(type) ->
@@ -99,10 +107,7 @@ defmodule Zoi.Types.Object do
       end)
     end
 
-    ## TODO: map_fetch infering string/atom keys should be done by explicitly passing a param
-    ## something like `:infer_keys` or `:coerce`
-    ## Add tests for input_map being string keys and key being an atom
-    defp map_fetch(input_map, key) do
+    defp map_fetch(input_map, key, _coerce = true) do
       Enum.map(input_map, fn {k, v} ->
         {to_string(k), v}
       end)
@@ -116,6 +121,10 @@ defmodule Zoi.Types.Object do
             Map.fetch(map, to_string(key))
         end
       end)
+    end
+
+    defp map_fetch(input_map, key, _coerce) do
+      Map.fetch(input_map, key)
     end
   end
 end